deleting duplicate foodtruck controllers file
